Refactor SDKv2 non-set detailed diff tests
