+++ pkg/tests/testdata/TestDetailedDiffList/list_attribute/list_element_removed_middle.golden
@@ -1,0 +1,52 @@
+tests.testOutput{
+	initialValue: &[]string{
+		"val1",
+		"val2",
+		"val3",
+	},
+	changeValue: &[]string{
+		"val3",
+		"val1",
+	},
+	tfOut: `
+Terraform used the selected providers to generate the following execution
+plan. Resource actions are indicated with the following symbols:
+  ~ update in-place
+
+Terraform will perform the following actions:
+
+  # crossprovider_test_res.example will be updated in-place
+  ~ resource "crossprovider_test_res" "example" {
+        id        = "newid"
+      ~ list_attr = [
+          - "val1",
+          - "val2",
+            "val3",
+          + "val1",
+        ]
+    }
+
+Plan: 0 to add, 1 to change, 0 to destroy.
+
+`,
+	pulumiOut: `Previewing update (test):
+  pulumi:pulumi:Stack: (same)
+    [urn=urn:pulumi:test::project::pulumi:pulumi:Stack::project-test]
+    ~ crossprovider:index/testRes:TestRes: (update)
+        [id=newid]
+        [urn=urn:pulumi:test::project::crossprovider:index/testRes:TestRes::example]
+      ~ listAttrs: [
+          ~ [0]: "val1" => "val3"
+          ~ [1]: "val2" => "val1"
+          - [2]: "val3"
+        ]
+Resources:
+    ~ 1 to update
+    1 unchanged
+`,
+	detailedDiff: map[string]interface{}{
+		"listAttrs[0]": map[string]interface{}{"kind": "UPDATE"},
+		"listAttrs[1]": map[string]interface{}{"kind": "UPDATE"},
+		"listAttrs[2]": map[string]interface{}{"kind": "DELETE"},
+	},
+}
