+++ pkg/tests/testdata/TestDetailedDiffList/list_block/list_element_removed_middle.golden
@@ -1,0 +1,62 @@
+tests.testOutput{
+	initialValue: &[]string{
+		"val1",
+		"val2",
+		"val3",
+	},
+	changeValue: &[]string{
+		"val3",
+		"val1",
+	},
+	tfOut: `
+Terraform used the selected providers to generate the following execution
+plan. Resource actions are indicated with the following symbols:
+  ~ update in-place
+
+Terraform will perform the following actions:
+
+  # crossprovider_test_res.example will be updated in-place
+  ~ resource "crossprovider_test_res" "example" {
+        id = "newid"
+
+      ~ list_block {
+          ~ prop = "val1" -> "val3"
+        }
+      ~ list_block {
+          ~ prop = "val2" -> "val1"
+        }
+      - list_block {
+          - prop = "val3" -> null
+        }
+    }
+
+Plan: 0 to add, 1 to change, 0 to destroy.
+
+`,
+	pulumiOut: `Previewing update (test):
+  pulumi:pulumi:Stack: (same)
+    [urn=urn:pulumi:test::project::pulumi:pulumi:Stack::project-test]
+    ~ crossprovider:index/testRes:TestRes: (update)
+        [id=newid]
+        [urn=urn:pulumi:test::project::crossprovider:index/testRes:TestRes::example]
+      ~ listBlocks: [
+          ~ [0]: {
+                  ~ prop: "val1" => "val3"
+                }
+          ~ [1]: {
+                  ~ prop: "val2" => "val1"
+                }
+          - [2]: {
+                  - prop: "val3"
+                }
+        ]
+Resources:
+    ~ 1 to update
+    1 unchanged
+`,
+	detailedDiff: map[string]interface{}{
+		"listBlocks[0].prop": map[string]interface{}{"kind": "UPDATE"},
+		"listBlocks[1].prop": map[string]interface{}{"kind": "UPDATE"},
+		"listBlocks[2]":      map[string]interface{}{"kind": "DELETE"},
+	},
+}
