+++ pkg/tests/testdata/TestDetailedDiffMap/key_changed.golden
@@ -1,0 +1,43 @@
+tests.testOutput{
+	initialValue: map[string]string{
+		"key": "val",
+	},
+	changeValue: map[string]string{"key2": "val"},
+	tfOut: `
+Terraform used the selected providers to generate the following execution
+plan. Resource actions are indicated with the following symbols:
+  ~ update in-place
+
+Terraform will perform the following actions:
+
+  # crossprovider_test_res.example will be updated in-place
+  ~ resource "crossprovider_test_res" "example" {
+        id       = "newid"
+      ~ map_prop = {
+          - "key"  = "val" -> null
+          + "key2" = "val"
+        }
+    }
+
+Plan: 0 to add, 1 to change, 0 to destroy.
+
+`,
+	pulumiOut: `Previewing update (test):
+  pulumi:pulumi:Stack: (same)
+    [urn=urn:pulumi:test::project::pulumi:pulumi:Stack::project-test]
+    ~ crossprovider:index/testRes:TestRes: (update)
+        [id=newid]
+        [urn=urn:pulumi:test::project::crossprovider:index/testRes:TestRes::example]
+      ~ mapProp: {
+          - key : "val"
+          + key2: "val"
+        }
+Resources:
+    ~ 1 to update
+    1 unchanged
+`,
+	detailedDiff: map[string]interface{}{
+		"mapProp.key":  map[string]interface{}{"kind": "DELETE"},
+		"mapProp.key2": map[string]interface{}{},
+	},
+}
