+++ pkg/tests/testdata/TestDetailedDiffString/changed.golden
@@ -1,0 +1,31 @@
+tests.testOutput{
+	initialValue: valast.Ptr("val1"), changeValue: valast.Ptr("val2"),
+	tfOut: `
+Terraform used the selected providers to generate the following execution
+plan. Resource actions are indicated with the following symbols:
+  ~ update in-place
+
+Terraform will perform the following actions:
+
+  # crossprovider_test_res.example will be updated in-place
+  ~ resource "crossprovider_test_res" "example" {
+        id          = "newid"
+      ~ string_prop = "val1" -> "val2"
+    }
+
+Plan: 0 to add, 1 to change, 0 to destroy.
+
+`,
+	pulumiOut: `Previewing update (test):
+  pulumi:pulumi:Stack: (same)
+    [urn=urn:pulumi:test::project::pulumi:pulumi:Stack::project-test]
+    ~ crossprovider:index/testRes:TestRes: (update)
+        [id=newid]
+        [urn=urn:pulumi:test::project::crossprovider:index/testRes:TestRes::example]
+      ~ stringProp: "val1" => "val2"
+Resources:
+    ~ 1 to update
+    1 unchanged
+`,
+	detailedDiff: map[string]interface{}{"stringProp": map[string]interface{}{"kind": "UPDATE"}},
+}
