+++ pkg/tests/testdata/TestDetailedDiffString/unchanged_non-empty.golden
@@ -1,0 +1,15 @@
+tests.testOutput{
+	initialValue: valast.Ptr("val1"), changeValue: valast.Ptr("val1"),
+	tfOut: `
+No changes. Your infrastructure matches the configuration.
+
+Terraform has compared your real infrastructure against your configuration
+and found no differences, so no changes are needed.
+`,
+	pulumiOut: `Previewing update (test):
+  pulumi:pulumi:Stack: (same)
+    [urn=urn:pulumi:test::project::pulumi:pulumi:Stack::project-test]
+Resources:
+    2 unchanged
+`,
+}
